feat(frontend): main layout improvements
